Create a class for nLinkReacherEnv that includes everything that is shared across the different environments
